+++ src/lib.rs
@@ -36,6 +36,16 @@ pub type NodeId = u64;
 /// to do a preliminary deserialization.
 pub trait AppData: Clone + Debug + Send + Sync + Serialize + DeserializeOwned + 'static {}
 
+/// A trait defining application specific response data.
+///
+/// The intention of this trait is that applications which are using this crate will be able to
+/// use their own concrete data types for returning response data from the storage layer when an
+/// entry is successfully applied to the state machine as part of a client request (this is not
+/// used during replication). This allows applications to seamlessly return application specific
+/// data from their storage layer, up through Raft, and back into their application for returning
+/// data to clients or other such uses.
+pub trait AppDataResponse: Clone + Debug + Send + Sync + Serialize + DeserializeOwned + 'static {}
+
 /// A trait defining application specific error types.
 ///
 /// The intention of this trait is that applications which are using this crate will be able to
@@ -44,8 +54,8 @@ pub trait AppData: Clone + Debug + Send + Sync + Serialize + DeserializeOwned + 
 /// to be able to communicate application specific logic from the storage layer.
 ///
 /// **NOTE WELL:** if an `AppError` is returned from any of the `RaftStorage` interfaces, other
-/// than the `AppendLogEntry` interface, then the Raft node will immediately shutdown. This is due
-/// to the fact that custom error handling logic is only allowed in the `AppendLogEntry` interface
+/// than the `AppendEntryToLog` interface, then the Raft node will immediately shutdown. This is due
+/// to the fact that custom error handling logic is only allowed in the `AppendEntryToLog` interface
 /// while the Raft node is the cluster leader. When the node is in any other state, the storage
 /// layer is expected to operate without any errors. Shutting down immediately is how Raft
 /// attempts to guard against data corruption and the like.